+++ Sources/KeyboardKit/Actions/KeyboardAction+RowItem.swift
@@ -13,5 +13,9 @@ import Foundation
  */
 extension KeyboardAction: RowItem {
 
+    /**
+     The row-specific ID to use when the action is presented
+     in a keyboard row.
+     */
     public var rowId: KeyboardAction { self }
 }